Extract room directory results to its own component (#8252)
