+++ src/i18n/strings/en_EN.json
@@ -1732,6 +1732,10 @@
     "Idle": "Idle",
     "Offline": "Offline",
     "Unknown": "Unknown",
+    "Unnamed room": "Unnamed room",
+    "Preview": "Preview",
+    "View": "View",
+    "Join": "Join",
     "Seen by %(userName)s at %(dateTime)s": "Seen by %(userName)s at %(dateTime)s",
     "Seen by %(displayName)s (%(userName)s) at %(dateTime)s": "Seen by %(displayName)s (%(userName)s) at %(dateTime)s",
     "Recently viewed": "Recently viewed",
@@ -1739,10 +1743,6 @@
     "Room %(name)s": "Room %(name)s",
     "Recently visited rooms": "Recently visited rooms",
     "No recently visited rooms": "No recently visited rooms",
-    "No rooms to show": "No rooms to show",
-    "Unnamed room": "Unnamed room",
-    "World readable": "World readable",
-    "Guests can join": "Guests can join",
     "(~%(count)s results)|other": "(~%(count)s results)",
     "(~%(count)s results)|one": "(~%(count)s result)",
     "Join Room": "Join Room",
@@ -2211,7 +2211,6 @@
     "Application window": "Application window",
     "Share content": "Share content",
     "Backspace": "Backspace",
-    "Join": "Join",
     "Please <newIssueLink>create a new issue</newIssueLink> on GitHub so that we can investigate this bug.": "Please <newIssueLink>create a new issue</newIssueLink> on GitHub so that we can investigate this bug.",
     "Something went wrong!": "Something went wrong!",
     "%(nameList)s %(transitionList)s": "%(nameList)s %(transitionList)s",
@@ -3021,8 +3020,6 @@
     "Couldn't find a matching Matrix room": "Couldn't find a matching Matrix room",
     "Fetching third party location failed": "Fetching third party location failed",
     "Unable to look up room ID from server": "Unable to look up room ID from server",
-    "Preview": "Preview",
-    "View": "View",
     "Create new room": "Create new room",
     "No results for \"%(query)s\"": "No results for \"%(query)s\"",
     "Try different words or check for typos. Some results may not be visible as they're private and you need an invite to join them.": "Try different words or check for typos. Some results may not be visible as they're private and you need an invite to join them.",
